Move OpenAiService to Lib.Services
