+++ src/Lib.Services/Services/OpenAiService/OpenAiServiceExtensions.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace MuzakBot.App.Services;
+namespace MuzakBot.Lib.Services;
 
 /// <summary>
 /// Extension methods for adding the OpenAI service to the service collection.
